LPD-69638 refactor: move site/site-ai-hub-rest-impl to ai-hub/ai-hub-rest-impl
