+++ build.properties
@@ -316,7 +316,6 @@
     #    dxp/apps/segments,\
     #    dxp/apps/sharepoint-rest,\
     #    dxp/apps/sharepoint-soap,\
-    #    dxp/apps/site,\
     #    util/portal-tools-db-migration-importer,\
     #    util/portal-tools-db-partition-migration-validator,\
     #    util/portal-tools-db-upgrade-client
