+++ modules/dxp/apps/ai-hub/ai-hub-rest-impl/bnd.bnd
@@ -1,0 +1,3 @@
+Bundle-Name: Liferay AI Hub REST Implementation
+Bundle-SymbolicName: com.liferay.ai.hub.rest.impl
+Bundle-Version: 1.0.0
