+++ modules/dxp/apps/ai-hub/ai-hub-rest-impl/rest-config.yaml
@@ -1,0 +1,12 @@
+apiDir: ../ai-hub-rest-api/src/main/java
+apiPackagePath: com.liferay.ai.hub.rest
+application:
+    baseURI: /ai-hub
+    className: AIHubRESTApplication
+    name: Liferay.AI.Hub.REST
+author: Feliphe Marinho
+clientDir: ../ai-hub-rest-client/src/main/java
+compatibilityVersion: 13
+forcePredictableOperationId: true
+javaEEPackage: jakarta
+testDir: ../ai-hub-rest-test/src/testIntegration/java
