--- modules/dxp/apps/site/site-ai-hub-rest-impl/bnd.bnd
@@ -1,3 +1,0 @@
-Bundle-Name: Liferay Site AI Hub REST Implementation
-Bundle-SymbolicName: com.liferay.site.ai.hub.rest.impl
-Bundle-Version: 1.0.0
