--- modules/dxp/apps/site/site-ai-hub-rest-impl/rest-openapi.yaml
@@ -1,47 +1,0 @@
-components:
-    schemas:
-        Task:
-            properties:
-                context:
-                    additionalProperties: true
-                    type: object
-                    writeOnly: true
-                externalReferenceCode:
-                    readOnly: true
-                    type: string
-                type:
-                    type: string
-            type: object
-info:
-    description:
-        "A Java client JAR is available for use with the group ID 'com.liferay', artifact ID
-        'com.liferay.notification.rest.client', and version '1.0.38'.. A Java client JAR is available for use with the
-        group ID 'com.liferay', artifact ID 'com.liferay.site.ai.hub.rest.client', and version '1.0.0'."
-    license:
-        name: Apache 2.0
-        url: http://www.apache.org/licenses/LICENSE-2.0.html
-    title: ""
-    version: v1.0
-openapi: 3.0.1
-paths:
-    "/tasks":
-        post:
-            operationId: postTask
-            requestBody:
-                content:
-                    application/json:
-                        schema:
-                            $ref: "#/components/schemas/Task"
-                    application/xml:
-                        schema:
-                            $ref: "#/components/schemas/Task"
-            responses:
-                200:
-                    content:
-                        application/json:
-                            schema:
-                                $ref: "#/components/schemas/Task"
-                        application/xml:
-                            schema:
-                                $ref: "#/components/schemas/Task"
-            tags: ["Task"]
