--- modules/dxp/apps/site/site-ai-hub-rest-impl/src/main/java/com/liferay/site/ai/hub/rest/internal/graphql/servlet/v1_0/ServletDataImpl.java
@@ -1,90 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2025 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.site.ai.hub.rest.internal.graphql.servlet.v1_0;
-
-import com.liferay.portal.kernel.util.ObjectValuePair;
-import com.liferay.portal.vulcan.graphql.servlet.ServletData;
-import com.liferay.site.ai.hub.rest.internal.graphql.mutation.v1_0.Mutation;
-import com.liferay.site.ai.hub.rest.internal.graphql.query.v1_0.Query;
-import com.liferay.site.ai.hub.rest.internal.resource.v1_0.TaskResourceImpl;
-import com.liferay.ai.hub.rest.resource.v1_0.TaskResource;
-
-import jakarta.annotation.Generated;
-
-import java.util.HashMap;
-import java.util.Map;
-
-import org.osgi.framework.BundleContext;
-import org.osgi.service.component.ComponentServiceObjects;
-import org.osgi.service.component.annotations.Activate;
-import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
-import org.osgi.service.component.annotations.ReferenceScope;
-
-/**
- * @author Feliphe Marinho
- * @generated
- */
-@Component(service = ServletData.class)
-@Generated("")
-public class ServletDataImpl implements ServletData {
-
-	@Activate
-	public void activate(BundleContext bundleContext) {
-		Mutation.setTaskResourceComponentServiceObjects(
-			_taskResourceComponentServiceObjects);
-	}
-
-	public String getApplicationName() {
-		return "Liferay.Site.AI.Hub.REST";
-	}
-
-	@Override
-	public Mutation getMutation() {
-		return new Mutation();
-	}
-
-	@Override
-	public String getPath() {
-		return "/ai-hub-graphql/v1_0";
-	}
-
-	@Override
-	public Query getQuery() {
-		return new Query();
-	}
-
-	public ObjectValuePair<Class<?>, String> getResourceMethodObjectValuePair(
-		String methodName, boolean mutation) {
-
-		if (mutation) {
-			return _resourceMethodObjectValuePairs.get(
-				"mutation#" + methodName);
-		}
-
-		return _resourceMethodObjectValuePairs.get("query#" + methodName);
-	}
-
-	private static final Map<String, ObjectValuePair<Class<?>, String>>
-		_resourceMethodObjectValuePairs =
-			new HashMap<String, ObjectValuePair<Class<?>, String>>() {
-				{
-					put(
-						"mutation#createTask",
-						new ObjectValuePair<>(
-							TaskResourceImpl.class, "postTask"));
-					put(
-						"mutation#createTaskBatch",
-						new ObjectValuePair<>(
-							TaskResourceImpl.class, "postTaskBatch"));
-				}
-			};
-
-	@Reference(scope = ReferenceScope.PROTOTYPE_REQUIRED)
-	private ComponentServiceObjects<TaskResource>
-		_taskResourceComponentServiceObjects;
-
-}
