--- modules/dxp/apps/site/site-ai-hub-rest-impl/src/main/resources/OSGI-INF/liferay/rest/v1_0/openapi.properties
@@ -1,9 +1,0 @@
-#
-# This is a generated file.
-#
-
-api.version=v1.0
-openapi.resource=true
-openapi.resource.path=/ai-hub
-osgi.jaxrs.application.select=(osgi.jaxrs.name=Liferay.Site.AI.Hub.REST)
-osgi.jaxrs.resource=true
