--- modules/dxp/apps/site/site-ai-hub-rest-impl/src/main/resources/OSGI-INF/liferay/rest/v1_0/task.properties
@@ -1,12 +1,0 @@
-#
-# This is a generated file.
-#
-
-api.version=v1.0
-batch.engine.entity.class.name=com.liferay.ai.hub.rest.dto.v1_0.Task
-batch.engine.task.item.delegate=true
-batch.planner.export.enabled=false
-batch.planner.import.enabled=true
-entity.class.name=com.liferay.ai.hub.rest.dto.v1_0.Task
-osgi.jaxrs.application.select=(osgi.jaxrs.name=Liferay.Site.AI.Hub.REST)
-osgi.jaxrs.resource=true
